refactor(loader): refactor the loader to top-level module
Updates the `build` script to store the revision hash in the `info.json` file instead of the `hash` property.

In the `en` translation file, updates the description key of the `action|{enable: ...}` translation to correctly indicate that enabling safe mode allows only the core plugins to run.

In the `metro/internal/modules.ts` file, updates the conditional statement to correctly handle the `cleanup` case.

In the `plugins/_api/settings/index.tsx` file, adds some logging for some patches.